fix: Update build-system.build-backend
refs https://python-poetry.org/docs/pyproject#poetry-and-pep-517

Fixes an error encountered when trying to add a local version of this package as an editable dependency to another local project:

```
poetry add -e <path-to-repo>
...

 ChefBuildError

  Backend operation failed: HookMissing('build_editable')

  at ~/.pyenv/versions/3.11.4/lib/python3.11/site-packages/poetry/installation/chef.py:147 in _prepare
      143│
      144│                 error = ChefBuildError("\n\n".join(message_parts))
      145│
      146│             if error is not None:
    → 147│                 raise error from None
      148│
      149│             return path
      150│
      151│     def _prepare_sdist(self, archive: Path, destination: Path | None = None) -> Path:

Note: This error originates from the build backend, and is likely not a problem with poetry but with strawberry-django-auth (0.376.4 ~/repos/strawberry-django-auth) not supporting PEP 517 builds. You can verify this by running 'pip wheel --use-pep517 --editable "/Users/jwegner/Data/development/repos/jacobwegner/strawberry-django-auth"'.
```

refs https://github.com/python-poetry/poetry/issues/7583#issuecomment-1453161741